(Development Status: Stable. Refactor: Login and Dashboard now flows together with sessions. Moved: PollRepository into repositories folder, userDashboard.html to users folder. Renamed: userDashboard.html to dashboard.html.
