--- src/main/resources/templates/users/protected.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Logout</title>
-</head>
-
-<body>
-    <a href="/logout">Logout</a>
-</body>
-
-</html>
